backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
slowapi==0.1.9

pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6

python-jose[cryptography]==3.3.0
bcrypt==5.0.0

pymongo==4.6.1
motor==3.3.2

httpx==0.26.0
redis==5.0.1

pytest==7.4.4
pytest-asyncio==0.23.3

email-validator

# ML Dependencies
transformers
torch
textblob

# Video Analysis
opencv-python==4.12.0.88
mediapipe==0.10.9
numpy==2.2.6

# Audio Analysis (Whisper + Audio Processing)
openai-whisper==20231117
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
ffmpeg-python==0.2.0

# PDF Generation
reportlab==4.4.7

# LLM/AI (Groq - free, fast)
# httpx already installed above
